% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-package.R
\docType{package}
\name{assert-package}
\alias{assert-package}
\alias{_PACKAGE}
\title{assert: Tidy assertions}
\description{
A re-imagining of stopifnot() with rlang::abort().
    Provides an api for simple input (and output) validation with consistent error messages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qfes/assert}
  \item Report bugs at \url{https://github.com/qfes/assert/issues}
}

}
\author{
\strong{Maintainer}: Anthony North \email{anthony.north@qfes.qld.gov.au}

Other contributors:
\itemize{
  \item Queensland Fire and Emergency Services [copyright holder, funder]
}

}
\keyword{internal}
