% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-assertions.R
\name{assert_equal}
\alias{assert_equal}
\title{Assert equal}
\usage{
assert_equal(a, b, error_message = "{a} must equal {b}", error_class = NULL)
}
\arguments{
\item{a}{<\code{any}> any value}

\item{b}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{a}} and \code{{b}}, which will be replaced with the unevaluated expressions
for \code{a} and \code{b}.}
}
\description{
Raises an assertion error when \code{!all(a == b)}.
}
\seealso{
Other assertions: 
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_not_null}()},
\code{\link{assert_range}()},
\code{\link{assert_type}()}
}
\concept{assertions}
