% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-assertions.R
\name{assert_has_names}
\alias{assert_has_names}
\title{Assert has names}
\usage{
assert_has_names(
  obj,
  names,
  error_message = "{obj} must have names {!!names}",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{obj}} and \code{{names}}, which will be replaced with the
unevaluated expressions for \code{obj} and \code{names}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!all(rlang::has_name(obj, names))}.
}
\seealso{
Other attribute-assertions: 
\code{\link{assert_is_named}()}
}
\concept{attribute-assertions}
