% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-assertions.R
\name{assert_inherits}
\alias{assert_inherits}
\title{Assert inherits}
\usage{
assert_inherits(
  obj,
  class,
  error_message = "{obj} must inherit {!!class}",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{class}{<\code{string} | \code{character}> the expected class(es)}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{obj}} and \code{{class}}, which will be replaced with the
unevaluated expressions for \code{obj} and \code{class}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!inherits(obj, class)}.
}
\seealso{
Other type-assertions: 
\code{\link{assert_is_character}()},
\code{\link{assert_is_double}()},
\code{\link{assert_is_integerish}()},
\code{\link{assert_is_integer}()},
\code{\link{assert_is_logical}()},
\code{\link{assert_is_numeric}()},
\code{\link{assert_is_raw}()},
\code{\link{assert_is_typeof}()}
}
\concept{type-assertions}
