% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-assertions.R
\name{assert_is_double}
\alias{assert_is_double}
\title{Assert is double}
\usage{
assert_is_double(
  obj,
  error_message = "{obj} must be a <double> vector",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholder \code{{obj}} which will be replaced with the unevaluated expression for \code{obj}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!rlang::is_double(obj)}.
}
\seealso{
Other type-assertions: 
\code{\link{assert_inherits}()},
\code{\link{assert_is_character}()},
\code{\link{assert_is_integerish}()},
\code{\link{assert_is_integer}()},
\code{\link{assert_is_logical}()},
\code{\link{assert_is_numeric}()},
\code{\link{assert_is_raw}()},
\code{\link{assert_is_typeof}()}
}
\concept{type-assertions}
