% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute-assertions.R
\name{assert_is_named}
\alias{assert_is_named}
\title{Assert is named}
\usage{
assert_is_named(
  obj,
  error_message = "{.arg obj} must be named",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{error_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!rlang::is_named(obj)}.
}
\seealso{
Other attribute-assertions: 
\code{\link{assert_has_names}()}
}
\concept{attribute-assertions}
