% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar-assertions.R
\name{assert_is_scalar_numeric}
\alias{assert_is_scalar_numeric}
\title{Assert is scalar numeric}
\usage{
assert_is_scalar_numeric(
  obj,
  error_message = "{obj} must be a scalar <numeric>",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholder \code{{obj}} which will be replaced with the unevaluated expression for \code{obj}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!is.numeric(obj) || length(obj) != 1L}.
}
\seealso{
Other scalar-assertions: 
\code{\link{assert_is_bool}()},
\code{\link{assert_is_scalar_character}()},
\code{\link{assert_is_scalar_double}()},
\code{\link{assert_is_scalar_integerish}()},
\code{\link{assert_is_scalar_integer}()},
\code{\link{assert_is_scalar_logical}()},
\code{\link{assert_is_scalar_raw}()},
\code{\link{assert_is_string}()}
}
\concept{scalar-assertions}
