% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-assertions.R
\name{assert_less}
\alias{assert_less}
\title{Assert less}
\usage{
assert_less(
  a,
  b,
  error_message = "{.arg a} must be less than {.arg b}",
  error_class = NULL
)
}
\arguments{
\item{a}{<\code{any}> any value}

\item{b}{<\code{any}> any value}

\item{error_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!all(a < b)}
}
\seealso{
Other logical-assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_range}()}
}
\concept{logical-assertions}
