% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-assertions.R
\name{assert_not_na}
\alias{assert_not_na}
\title{Assert not na}
\usage{
assert_not_na(
  obj,
  error_message = "{.arg obj} must be not contain {.code NA}",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any object}

\item{error_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{anyNA(obj)}
}
\seealso{
Other assertions: 
\code{\link{assert_not_null}()}
}
\concept{assertions}
