% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-assertions.R
\name{assert_type}
\alias{assert_type}
\title{Assert type}
\usage{
assert_type(
  obj,
  type,
  error_message = c(x = "{obj} must be type {!!type}"),
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{type}{<\code{string}> the expected type}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{obj}} and \code{{type}}, which will be replaced with the
unevaluated expressions for \code{obj} and \code{type}.}
}
\description{
Raises an assertion error when \code{typeof(obj) != type}.
}
\seealso{
Other assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_not_null}()},
\code{\link{assert_range}()}
}
\concept{assertions}
