% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar-assertions.R
\name{assert_is_scalar_logical}
\alias{assert_is_scalar_logical}
\title{Assert is scalar logical}
\usage{
assert_is_scalar_logical(
  obj,
  error_message = "{obj} must be a scalar <logical>",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholder \code{{obj}} which will be replaced with the unevaluated expression for \code{obj}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!rlang::is_scalar_logical(obj)}.
}
\seealso{
Other assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_is_bool}()},
\code{\link{assert_is_scalar_character}()},
\code{\link{assert_is_scalar_double}()},
\code{\link{assert_is_scalar_integerish}()},
\code{\link{assert_is_scalar_integer}()},
\code{\link{assert_is_scalar_raw}()},
\code{\link{assert_is_string}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_not_null}()},
\code{\link{assert_range}()},
\code{\link{assert_typeof}()}
}
\concept{assertions}
