% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-assertions.R
\name{assert_is_typeof}
\alias{assert_is_typeof}
\title{Assert typeof}
\usage{
assert_is_typeof(
  obj,
  type,
  error_message = "{.arg obj} must be of type {.cls {type}}",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any value}

\item{type}{<\code{string}> the expected type}

\item{error_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{typeof(obj) != type}.
}
\seealso{
Other type-assertions: 
\code{\link{assert_inherits}()},
\code{\link{assert_is_character}()},
\code{\link{assert_is_double}()},
\code{\link{assert_is_integerish}()},
\code{\link{assert_is_integer}()},
\code{\link{assert_is_logical}()},
\code{\link{assert_is_numeric}()},
\code{\link{assert_is_raw}()}
}
\concept{type-assertions}
