% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-assertions.R
\name{assert_less}
\alias{assert_less}
\title{Assert less}
\usage{
assert_less(
  a,
  b,
  error_message = "{a} must be less than {b}",
  error_class = NULL
)
}
\arguments{
\item{a}{<\code{any}> any value}

\item{b}{<\code{any}> any value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{a}} and \code{{b}}, which will be replaced with the unevaluated expressions
for \code{a} and \code{b}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!all(a < b)}
}
\seealso{
Other assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_is_bool}()},
\code{\link{assert_is_scalar_character}()},
\code{\link{assert_is_scalar_double}()},
\code{\link{assert_is_scalar_integer}()},
\code{\link{assert_is_scalar_logical}()},
\code{\link{assert_is_scalar_raw}()},
\code{\link{assert_is_string}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_not_null}()},
\code{\link{assert_range}()},
\code{\link{assert_typeof}()}
}
\concept{assertions}
