% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-assertions.R
\name{assert_not_na}
\alias{assert_not_na}
\title{Assert not na}
\usage{
assert_not_na(
  obj,
  error_message = "{obj} must be not contain NA",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any object}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{obj}} which will be replaced with the unevaluated expression
for \code{obj}.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{anyNA(obj)}
}
\seealso{
Other assertions: 
\code{\link{assert_not_null}()}
}
\concept{assertions}
