% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-assertions.R
\name{assert_not_null}
\alias{assert_not_null}
\title{Assert not null}
\usage{
assert_not_null(
  obj,
  error_message = "{obj} must be not NULL",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any object}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{obj}} which will be replaced with the unevaluated expression
for \code{obj}.}
}
\description{
Raises an assertion error when \code{is.null(obj)}
}
\seealso{
Other assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_range}()},
\code{\link{assert_type}()}
}
\concept{assertions}
