% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-assertions.R
\name{assert_range}
\alias{assert_range}
\title{Assert range}
\usage{
assert_range(
  obj,
  min,
  max,
  error_message = "{obj} must be within [{!!min}, {!!max}]",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any object}

\item{min}{<\code{any}> the minimum value}

\item{max}{<\code{any}> the maximum value}

\item{error_message}{<\code{string}> the error message.
Accepts placeholders \code{{a}} and \code{{b}}, which will be replaced with the unevaluated expressions
for \code{a} and \code{b}.}
}
\description{
Raises an assertion error when \code{!all(a >= min & a <= max)}
}
\seealso{
Other assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_inherits}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()},
\code{\link{assert_not_na}()},
\code{\link{assert_not_null}()},
\code{\link{assert_typeof}()}
}
\concept{assertions}
