% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-assertions.R
\name{assert_range}
\alias{assert_range}
\title{Assert range}
\usage{
assert_range(
  obj,
  min,
  max,
  error_message = "{.arg obj} must be within [{.arg min}, {.arg max}]",
  error_class = NULL
)
}
\arguments{
\item{obj}{<\code{any}> any object}

\item{min}{<\code{any}> the minimum value}

\item{max}{<\code{any}> the maximum value}

\item{error_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{error_class}{<\code{character}> the class name/s for the error.}
}
\description{
Raises an assertion error when \code{!all(a >= min & a <= max)}
}
\seealso{
Other logical-assertions: 
\code{\link{assert_equal}()},
\code{\link{assert_greater_equal}()},
\code{\link{assert_greater}()},
\code{\link{assert_less_equal}()},
\code{\link{assert_less}()},
\code{\link{assert_not_equal}()}
}
\concept{logical-assertions}
