% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert-soft.R
\name{warn_if_not}
\alias{warn_if_not}
\alias{warn_if}
\title{Assert (soft)}
\usage{
warn_if_not(
  expr,
  warn_message = NULL,
  warn_class = NULL,
  env = rlang::caller_env(2L),
  print_expr = NULL,
  ...
)

warn_if(
  expr,
  warn_message = NULL,
  warn_class = NULL,
  env = rlang::caller_env(2L),
  print_expr = NULL,
  ...
)
}
\arguments{
\item{expr}{<\code{expression}> a logical expression to test.}

\item{warn_message}{<\code{string}> a message to be displayed when assertion fails.}

\item{warn_class}{<\code{character}> the class name/s for the warning.}

\item{env}{<\code{environment}> the environment for substituted \code{dots} and \code{print_expr}.
Has no effect if \code{dots} and \code{print_expr} are already quosures.}

\item{print_expr}{<\code{expression}> a diffused expression for altering the error message.
Defaults to \code{rlang::as_quosure(substitute(expr), rlang::caller_env())}}

\item{...}{<\code{any}> values used in evaluating glue expressions, for the error message.}
}
\description{
Raises an assertion warning when \code{any(expr)} is false.

Raises an assertion warning when \code{all(expr)} is true.
}
